Add ID steps to prevent overlapping db and cache auction IDs
The AH cache doesn't ever check if the IDs from auctions made in-game are correct and doesn't include/iterate over the auctions loaded from ".reload auctions", making it throw a db error on attempt to insert on an ID already inserted by the Eluna AH bot. 

This adds a gap between bot auctions and player auctions of 100 000 000.